Update EleRelPointPairPortable.h fix crashes and update collection --> Collection
