+++ DataFormats/EgammaReco/src/alpaka/classes_cuda_def.xml
@@ -1,7 +1,7 @@
 <lcgdict>
-  <class name="alpaka_cuda_async::reco::SuperclusterDeviceCollection" persistent="false"/>
-  <class name="edm::DeviceProduct<alpaka_cuda_async::reco::SuperclusterDeviceCollection>" persistent="false"/>
-  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_cuda_async::reco::SuperclusterDeviceCollection>>" persistent="false"/>
+  <class name="alpaka_cuda_async::reco::SuperClusterDeviceCollection" persistent="false"/>
+  <class name="edm::DeviceProduct<alpaka_cuda_async::reco::SuperClusterDeviceCollection>" persistent="false"/>
+  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_cuda_async::reco::SuperClusterDeviceCollection>>" persistent="false"/>
 
   <class name="alpaka_cuda_async::reco::EleSeedDeviceCollection" persistent="false"/>
   <class name="edm::DeviceProduct<alpaka_cuda_async::reco::EleSeedDeviceCollection>" persistent="false"/>
