+++ DataFormats/EgammaReco/src/alpaka/classes_rocm.h
@@ -4,7 +4,7 @@
 #include "DataFormats/Common/interface/Wrapper.h"
 
 #include "DataFormats/EgammaReco/interface/SuperClusterSoA.h"
-#include "DataFormats/EgammaReco/interface/alpaka/SuperclusterDeviceCollection.h"
+#include "DataFormats/EgammaReco/interface/alpaka/SuperClusterDeviceCollection.h"
 
 #include "DataFormats/EgammaReco/interface/EleSeedSoA.h"
 #include "DataFormats/EgammaReco/interface/alpaka/EleSeedDeviceCollection.h"
