+++ DataFormats/EgammaReco/src/alpaka/classes_rocm_def.xml
@@ -1,7 +1,7 @@
 <lcgdict>
-  <class name="alpaka_rocm_async::reco::SuperclusterDeviceCollection" persistent="false"/>
-  <class name="edm::DeviceProduct<alpaka_rocm_async::reco::SuperclusterDeviceCollection>" persistent="false"/>
-  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_rocm_async::reco::SuperclusterDeviceCollection>>" persistent="false"/>
+  <class name="alpaka_rocm_async::reco::SuperClusterDeviceCollection" persistent="false"/>
+  <class name="edm::DeviceProduct<alpaka_rocm_async::reco::SuperClusterDeviceCollection>" persistent="false"/>
+  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_rocm_async::reco::SuperClusterDeviceCollection>>" persistent="false"/>
 
   <class name="alpaka_rocm_async::reco::EleSeedDeviceCollection" persistent="false"/>
   <class name="edm::DeviceProduct<alpaka_rocm_async::reco::EleSeedDeviceCollection>" persistent="false"/>
