+++ RecoEgamma/EgammaElectronProducers/plugins/alpaka/PixelMatchingAlgo.h
@@ -3,7 +3,7 @@
 
 
 #include "DataFormats/EgammaReco/interface/alpaka/EleSeedDeviceCollection.h"
-#include "DataFormats/EgammaReco/interface/alpaka/SuperclusterDeviceCollection.h"
+#include "DataFormats/EgammaReco/interface/alpaka/SuperClusterDeviceCollection.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/config.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
@@ -11,8 +11,8 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
   class PixelMatchingAlgo {
   public:
     void printEleSeeds(Queue& queue, const reco::EleSeedDeviceCollection& collection) const;
-    void printSCs(Queue& queue, const reco::SuperclusterDeviceCollection& collection) const;
-    void matchSeeds(Queue& queue, reco::EleSeedDeviceCollection& collection, reco::SuperclusterDeviceCollection& collectionSCs, double vtx_X, double vtx_Y, double vtx_Z) const;
+    void printSCs(Queue& queue, const reco::SuperClusterDeviceCollection& collection) const;
+    void matchSeeds(Queue& queue, reco::EleSeedDeviceCollection& collection, reco::SuperClusterDeviceCollection& collectionSCs, double vtx_X, double vtx_Y, double vtx_Z) const;
 
   };
